Mege Developer funzionante in Main

+++ NuovoCinemaParadiso/Dtos/DtoCreazioneTipologiaSala.cs
@@ -1,10 +1,9 @@
 using System.ComponentModel.DataAnnotations;
 
 namespace NuovoCinemaParadiso.Dtos;
-
 public class DtoCreazioneTipologiaSala
 {
     [Required]
     public string Nome { get; set; } = string.Empty;
-    public decimal MaggiorazionePrezzo { get; set; }
+    public int MaggiorazionePrezzo { get; set; }
 }
